reporteEquipoLocust/Ejercicio_Pruebas_Locust/Libros/requirements.txt
Flask==3.0.3
gunicorn==21.2.0
locust==2.31.5
mysql-connector-python==9.0.0
redis==5.0.1
Faker==30.3.0
requests==2.32.3
